# contracts/media-control.openapi.yaml
openapi: 3.0.3
info:
  title: MIMIR Media Bridge Control API
  version: 2.0.0
  description: |
    Versioned HTTP contract between the orchestrator and the media bridge.
    This contract documents the provider-agnostic media API implemented in this repository today.

    Scope notes:
      - Inbound media only.
      - Session lifecycle operations are keyed by media session ID.
      - Media lifecycle events are delivered over server-sent events (SSE).
servers:
  - url: http://media-bridge:8081
paths:
  /v1/media/sessions:
    post:
      summary: Create a media session for an inbound call
      operationId: CreateMediaSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMediaSessionRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaSession'
        '403':
          description: Rejected because the request is not inbound
  /v1/media/sessions/{session_id}/start:
    post:
      summary: Start media for an existing session
      operationId: StartMediaSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Media session state after start
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaSession'
        '404':
          description: Session not found
  /v1/media/sessions/{session_id}/stop:
    post:
      summary: Stop media for an existing session
      operationId: StopMediaSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StopMediaSessionRequest'
      responses:
        '200':
          description: Media session state after stop
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaSession'
        '404':
          description: Session not found
  /v1/media/sessions/{session_id}:
    get:
      summary: Get media session state
      operationId: GetMediaSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Current media session state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaSession'
        '404':
          description: Session not found
  /v1/media/sessions/{session_id}/telemetry:
    post:
      summary: Record media runtime telemetry for a session
      operationId: PostSessionTelemetry
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryPayload'
      responses:
        '200':
          description: Telemetry accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        '404':
          description: Session not found
  /v1/media/events:
    get:
      summary: Stream media lifecycle events from the media bridge
      operationId: StreamMediaEvents
      parameters:
        - in: query
          name: session_id
          required: false
          schema:
            type: string
          description: Optional comma-separated list of media session IDs
        - in: query
          name: call_id
          required: false
          schema:
            type: string
          description: Optional call ID filter
      responses:
        '200':
          description: Server-sent event stream
          content:
            text/event-stream:
              schema:
                type: string
components:
  parameters:
    SessionId:
      in: path
      name: session_id
      required: true
      schema:
        type: string
      description: Media session identifier generated by the media bridge
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
      description: Idempotency token required by start and stop operations
  schemas:
    CallParticipant:
      type: object
      required: [caller, callee, called_extension]
      properties:
        caller:
          type: string
        callee:
          type: string
        called_extension:
          type: string
    RtpFlow:
      type: object
      required: [local_address, local_port, remote_address, remote_port]
      properties:
        local_address:
          type: string
        local_port:
          type: integer
        remote_address:
          type: string
        remote_port:
          type: integer
    AIProfile:
      type: object
      required: [model_name, voice, instructions, greeting, vad_mode]
      properties:
        model_name:
          type: string
        voice:
          type: string
        instructions:
          type: string
        greeting:
          type: string
        vad_mode:
          type: string
    MediaSettings:
      type: object
      properties:
        input_codec:
          type: string
          default: g711_ulaw
        output_codec:
          type: string
          default: g711_ulaw
        sample_rate_hz:
          type: integer
          default: 8000
    CreateMediaSessionRequest:
      type: object
      required: [call_id, direction, participant, ai_profile, rtp]
      properties:
        call_id:
          type: string
        direction:
          type: string
          enum: [inbound]
        participant:
          $ref: '#/components/schemas/CallParticipant'
        ai_profile:
          $ref: '#/components/schemas/AIProfile'
        media_settings:
          $ref: '#/components/schemas/MediaSettings'
        rtp:
          $ref: '#/components/schemas/RtpFlow'
        metadata:
          type: object
          additionalProperties:
            type: string
    StopMediaSessionRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    MediaSession:
      type: object
      required: [session_id, bridge_session_id, call_id, status]
      properties:
        session_id:
          type: string
        bridge_session_id:
          type: string
        call_id:
          type: string
        status:
          type: string
          enum: [created, active, terminated]
        reason:
          type: string
          nullable: true
    TelemetryPayload:
      type: object
      properties:
        runtime:
          type: string
          default: python
        packet_loss_pct:
          type: number
          format: float
          nullable: true
        jitter_ms:
          type: number
          format: float
          nullable: true
        ws_reconnects:
          type: integer
          default: 0
        ws_errors:
          type: integer
          default: 0
        first_audio_latency_ms:
          type: number
          format: float
          nullable: true
    AcceptedResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: accepted
    MediaLifecycleEvent:
      type: object
      required: [event_id, event_type, bridge_session_id, call_id, occurred_at, attributes]
      properties:
        event_id:
          type: string
        event_type:
          type: string
          example: media.session.active
        media_session_id:
          type: string
          nullable: true
        bridge_session_id:
          type: string
          nullable: true
        call_id:
          type: string
        occurred_at:
          type: string
          format: date-time
        attributes:
          type: object
          additionalProperties: true